[feat] Update public signature to permit config
In the future mokku is likely to support different options (e.g. whether
or not to include godocs on the generated type, make mocks safe for
concurrent access, whether or not to store the arguments per call to a
mocked method, etc.).
In order to not having to break backwards compatibility (even before
v1.x.x) this commit identifies the desired signature upfront.